Saving Strategy#client result in Strategy@client ivar, so multiple calls to Strategy#client won't create a new client instance each time.
Added client_options :history, which defaults to Array.new.
This will be used by the Client instance to store API request/response
history for the duration of the instance life cycle. The Client history
is used to populate the auth_hash['extra']['raw_info'] section.
If client_options :history is set to nil, Client history is disabled.